Refactor ReplaceOffer to reuse code for restore state.
