+++ tools/c7n_azure/c7n_azure/entry.py
@@ -25,6 +25,7 @@ import c7n_azure.container_host.modes
 import c7n_azure.resources.aci
 import c7n_azure.resources.generic_arm_resource
 import c7n_azure.resources.cosmos_db
+import c7n_azure.resources.cosmos_db_child_resources
 import c7n_azure.resources.key_vault
 import c7n_azure.resources.key_vault_keys
 import c7n_azure.resources.load_balancer
