+++ tools/c7n_azure/tests/cassettes/CosmosDBChildResourcesTest.test_offer_collection.json
