+++ tools/c7n_azure/tests/test_cosmos_db.py
@@ -131,80 +131,3 @@ class CosmosDBTest(BaseTest):
         }, validate=True)
         resources = p.run()
         self.assertEqual(0, len(resources))
-
-    @arm_template('cosmosdb.json')
-    def test_offer_collection(self):
-        p = self.load_policy({
-            'name': 'test-azure-cosmosdb',
-            'resource': 'azure.cosmosdb-collection',
-            'filters': [
-                {'type': 'offer',
-                 'key': 'content.offerThroughput',
-                 'op': 'gt',
-                 'value': 100}],
-        })
-        resources = p.run()
-
-        self.assertEqual(1, len(resources))
-        self.assertEqual('Hash', resources[0]['partitionKey']['kind'])
-
-
-class CosmosDBReplaceOfferActionTest(BaseTest):
-
-    def setUp(self, *args, **kwargs):
-        super(CosmosDBReplaceOfferActionTest, self).setUp(*args, **kwargs)
-        client = local_session(Session).client('azure.mgmt.cosmosdb.CosmosDB')
-        key = CosmosDBChildResource.get_cosmos_key(
-            'test_cosmosdb', 'cctestcosmosdb', client, readonly=False)
-        self.data_client = CosmosClient(
-            url_connection='https://cctestcosmosdb.documents.azure.com:443/',
-            auth={
-                'masterKey': key
-            }
-        )
-
-    def tearDown(self, *args, **kwargs):
-        super(CosmosDBReplaceOfferActionTest, self).tearDown(*args, **kwargs)
-        self.data_client.ReplaceOffer(
-            self.initial_offer['_self'],
-            self.initial_offer
-        )
-
-    def test_replace_offer_collection_action(self):
-
-        p = self.load_policy({
-            'name': 'test-azure-cosmosdb',
-            'resource': 'azure.cosmosdb-collection',
-            'filters': [
-                {
-                    'type': 'value',
-                    'key': 'id',
-                    'op': 'eq',
-                    'value': 'cccontainer'
-                },
-                {
-                    'type': 'offer',
-                    'key': 'content.offerThroughput',
-                    'op': 'eq',
-                    'value': 400
-                }
-            ],
-            'actions': [
-                {
-                    'type': 'replace-offer',
-                    'throughput': 500
-                }
-            ]
-        })
-        collections = p.run()
-        self.assertEqual(len(collections), 1)
-
-        self.initial_offer = collections[0]['c7n:offer'][0]
-        self.collection = collections[0]
-        self._assert_offer_throughput_equals(500, collections[0]['_self'])
-
-    def _assert_offer_throughput_equals(self, throughput, resource_self):
-        offers = self.data_client.ReadOffers()
-        offer = next((o for o in offers if o['resource'] == resource_self), None)
-        self.assertIsNotNone(offer)
-        self.assertEqual(offer['content']['offerThroughput'], throughput)
